fixed issue with deleting a product

+++ products/views.py
@@ -1,4 +1,4 @@
-from django.shortcuts import render, redirect, reverse, get_object_or_404
+from django.shortcuts import render, redirect, reverse, get_object_or_404, HttpResponseRedirect
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from django.db.models import Q
@@ -119,12 +119,25 @@ def delete_product(request, product_id):
     """ Delete a product from the store """
     if not request.user.is_superuser:
         messages.error(request, 'Sorry, only store owners can do that.')
-        return redirect(reverse('home'))
+        return redirect(reverse('home'))    
 
     product = get_object_or_404(Product, pk=product_id)
-    product.delete()
-    messages.success(request, 'Product deleted!')
-    return redirect(reverse('products'))
+
+    if request.method == "POST":
+        # delete object
+        product.delete()
+        messages.success(request, 'Product deleted!')
+        # after deleting redirect to
+        # products page
+        # return HttpResponseRedirect('products')
+        return redirect(reverse('products'))
+
+    template = 'products/delete_product.html'
+    context = {
+         'product': product,
+     }
+
+    return render(request, "products/delete_product.html", context)
 
 
 def product_detail(request, product_id):
